# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for querying time series data from TimescaleDB",
        "title": "Data Query Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/alerts": {
            "get": {
                "description": "Query alert events with optional filters (all parameters are optional). Time filter uses overlap logic: alert matches if its time range overlaps with query time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Query alerts with optional filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance endpoint filter",
                        "name": "endpoint",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Alert text keyword filter (case-insensitive LIKE match)",
                        "name": "alert_text",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time filter - alerts active after this time (RFC3339 or Unix timestamp)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time filter - alerts started before this time (RFC3339 or Unix timestamp)",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Metric name filter (exact match)",
                        "name": "metric",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status filter (exact match)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "maximum": 100,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.AlertsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/slow-queries": {
            "get": {
                "description": "Query slow SQL query records with optional filters (all parameters are optional). Supports filtering by endpoint, SQL keyword, and time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slow-queries"
                ],
                "summary": "Query slow queries with optional filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance endpoint filter",
                        "name": "endpoint",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SQL text keyword filter (case-insensitive LIKE match)",
                        "name": "sql_keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or Unix timestamp)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or Unix timestamp)",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "maximum": 100,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.SlowQueryListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/endpoints": {
            "get": {
                "description": "Get all distinct endpoints from the time series database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "Get all endpoints",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.EndpointsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instances": {
            "get": {
                "description": "Get all database instances with pagination support",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get all instances with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "maximum": 100,
                        "description": "Number of items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.InstancesListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/instances/{endpoint}": {
            "get": {
                "description": "Get database instance metadata (db_type, entity_name, instance_vip, instance_port, etc.) by endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get instance metadata by endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance endpoint",
                        "name": "endpoint",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.InstanceMetaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/series": {
            "get": {
                "description": "Query time series data with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Query series data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Endpoint filter",
                        "name": "endpoint",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Metric filter",
                        "name": "metric",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Label filter expression",
                        "name": "label_filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or Unix timestamp)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or Unix timestamp)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sampling interval (e.g., 5m, 1h). Returns averaged values within each time bucket. Valid range: 1m to 1h.",
                        "name": "step",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.SeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/series/query": {
            "post": {
                "description": "Query multiple time series with complex filters and optional aggregation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Query multiple series",
                "parameters": [
                    {
                        "description": "Query parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dataquery.SeriesQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.SeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/series/{id}": {
            "get": {
                "description": "Get a single time series by its ID with optional time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or Unix timestamp, default: 1 hour ago)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or Unix timestamp, default: now)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataquery.SeriesSingleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dataquery.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dataquery.Alert": {
            "type": "object",
            "properties": {
                "alert_text": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "metric": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dataquery.AlertsListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dataquery.Alert"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dataquery.PaginationMeta"
                }
            }
        },
...
        "dataquery.EndpointsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dataquery.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dataquery.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dataquery.ErrorDetail"
                }
            }
        },
        "dataquery.InstanceMeta": {
            "type": "object",
            "properties": {
                "alert_subscriber": {
                    "type": "string"
                },
                "attach_db": {
                    "type": "string"
                },
                "backup_method": {
                    "type": "string"
                },
                "business_owner": {
                    "type": "string"
                },
                "ccm_name": {
                    "type": "string"
                },
                "character_set": {
                    "type": "string"
                },
                "chinese_desc": {
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "db_type": {
                    "type": "string"
                },
                "default_role": {
                    "type": "string"
                },
                "entity_name": {
                    "type": "string"
                },
                "environment": {
                    "type": "string"
                },
                "failover_type": {
                    "type": "string"
                },
                "ha_type": {
                    "type": "string"
                },
                "host_environment1": {
                    "type": "string"
                },
                "host_environment2": {
                    "type": "string"
                },
                "host_infra_type1": {
                    "type": "string"
                },
                "host_infra_type2": {
                    "type": "string"
                },
                "host_ip1": {
                    "type": "string"
                },
                "host_ip2": {
                    "type": "string"
                },
                "host_name2": {
                    "type": "string"
                },
                "host_namel": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "infra_type": {
                    "type": "string"
                },
                "ins_created_date": {
                    "type": "string"
                },
                "ins_updated_date": {
                    "type": "string"
                },
                "ins_uuid": {
                    "type": "string"
                },
                "instance_endpoint": {
                    "type": "string"
                },
                "instance_name": {
                    "type": "string"
                },
                "instance_port": {
                    "type": "integer"
                },
                "instance_vip": {
                    "type": "string"
                },
                "is_created_by_cloud": {
                    "type": "string"
                },
                "le_name": {
                    "type": "string"
                },
                "opr_dba": {
                    "type": "string"
                },
                "opr_dba_ii": {
                    "type": "string"
                },
                "org_code": {
                    "type": "string"
                },
                "os_name": {
                    "type": "string"
                },
                "req_cpu": {
                    "type": "number"
                },
                "req_memory_gb": {
                    "type": "number"
                },
                "req_storage_gb": {
                    "type": "number"
                },
                "role": {
                    "type": "string"
                },
                "service_user": {
                    "type": "string"
                },
                "source_sys": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subsys_code": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "version_detail": {
                    "type": "string"
                }
            }
        },
        "dataquery.InstanceMetaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dataquery.InstanceMeta"
                }
            }
        },
        "dataquery.InstancesListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dataquery.InstanceMeta"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dataquery.PaginationMeta"
                }
            }
        },
        "dataquery.PaginationMeta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dataquery.SeriesDataDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "labels_hash": {
                    "type": "string"
                },
                "metric": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dataquery.DataPointDTO"
                    }
                },
                "statistics": {
                    "$ref": "#/definitions/dataquery.SeriesStatisticsDTO"
                }
            }
        },
        "dataquery.SeriesQueryRequestBody": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "endpoints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "labels": {
                    "type": "string"
                },
                "metrics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "dataquery.SeriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dataquery.SeriesDataDTO"
                    }
                }
            }
        },
        "dataquery.SeriesSingleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dataquery.SeriesDataDTO"
                }
            }
        },
...
        "dataquery.SlowQuery": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "endpoint": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "database_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "sql_text": {
                    "type": "string"
                },
                "execute_time": {
                    "type": "number"
                },
                "execute_date": {
                    "type": "string"
                }
            }
        },
        "dataquery.SlowQueryListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dataquery.SlowQuery"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dataquery.PaginationMeta"
                }
            }
        }
    }
}